+++ packages/wallet-cli/src/daemon/wallet-factory.test.ts
@@ -124,9 +124,7 @@ describe('createWallet', () => {
     );
     expect(instanceOptions.transactionController?.disableSwaps).toBe(true);
     expect(instanceOptions.transactionController?.hooks).toStrictEqual({});
-    expect(instanceOptions.shieldController.baseUrl).toBe(
-      'https://ruleset-engine.api.cx.metamask.io',
-    );
+    expect(instanceOptions.shieldController.env).toBeUndefined();
     expect(instanceOptions.shieldController.fetchFunction).toBe(
       globalThis.fetch,
     );