+++ packages/wallet/src/Wallet.test.ts
@@ -29,7 +29,6 @@ const REMOTE_FEATURE_FLAG_OPTIONS = {
 };
 
 const SHIELD_CONTROLLER_OPTIONS = {
-  baseUrl: 'https://ruleset-engine.api.cx.metamask.io',
   fetchFunction: jest.fn(),
   backend: {
     checkCoverage: jest.fn(),